requirements.txt
google-genai>=1.0.0
ccxt>=4.3.0
pandas>=2.0.0
numpy>=1.26.0
python-dotenv>=1.0.0
rich>=13.7.0
aiohttp>=3.9.0
requests>=2.31.0
